Update date to long

+++ core/src/main/kotlin/dev/gumil/nookbot/entities/telegram/Message.kt
@@ -13,7 +13,7 @@ data class Message(
     val from: User? = null,
 
     @SerialName("date")
-    val date: Int,
+    val date: Long,
 
     @SerialName("chat")
     val chat: Chat,
